v1/order/requirements.txt
flask
flask-sqlalchemy
mysqlclient
flask_script
requests
flask-debugtoolbar



